Added delete, update views and some updates for find Contact by ID. New alert messages are added

+++ frontend/src/app/contact-delete/contact-delete.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
-import { HttpClient, HttpHeaders } from '@angular/common/http';
+
+import { HttpClient } from '@angular/common/http';
+import { FormControl } from '@angular/forms';
 
 @Component({
   selector: 'app-contact-delete',
@@ -7,17 +9,28 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
   styleUrl: './contact-delete.component.css'
 })
 export class ContactDeleteComponent {
+  title = 'contact-delete';
 
   constructor(private http: HttpClient) { }
 
-  contactId: string = "1";
+  idValue = new FormControl(0);
 
   deleteContact(): void {
+    if(this.idValue.value == undefined || this.idValue.value == null || this.idValue.value === 0){
+      alert('Invalid ID!');
+      return;
+    }
     this.http.delete(
-      `http://localhost:5000/delete/${this.contactId}`, {withCredentials: true}
-    ).subscribe(data =>
-      //this.contacts = data
-      console.log(data)
+      'http://localhost:5000/delete/' + this.idValue.value, 
+      {observe: 'response', withCredentials: true}
+    ).subscribe((data) => {
+      if(data.ok){
+        alert('Contact deleted!');
+      }
+    }, 
+    (error) => {
+      alert('Contact not found!');
+    }
     );
   }
